feat: allow float alpha and beta3 warmup
E.g. 0.5 means half of an epoch for warmup

+++ train.py
@@ -85,10 +85,10 @@ def launch(
     model = TransformerSeq(
         vocab_size=data_params["vocab_size"],
         world_size = world_size,
-        gamma1_warmup = model_params["gamma_warmup"] * entries_per_epoch,
+        gamma1_warmup = int(model_params["gamma_warmup"] * entries_per_epoch),
         total_steps = trainer_params["epochs"] * entries_per_epoch,
-        alpha_warmup = model_params["alpha_warmup_ep"] * entries_per_epoch,
-        beta3_warmup = model_params["beta3_warmup_ep"] * entries_per_epoch,
+        alpha_warmup = int(model_params["alpha_warmup_ep"] * entries_per_epoch),
+        beta3_warmup = int(model_params["beta3_warmup_ep"] * entries_per_epoch),
         **model_params,
     )
     if is_master:
